src/python/requirements.txt
openai>=1.50.0
azure-identity>=1.15.0
